+++ README.md
@@ -1,13 +1,3 @@
-# RxID Pill Indentification Project (Lambda-Labs12)
-## Data Science Repository
-   Mark Oliver, Carlos Gutierrez, Ed Chin
+# DS Backend Module  setup for deployment to AWS
 
-DS Web Service endpoint: 
-http://rxid-ds.us-east-2.elasticbeanstalk.com/
-
-Database hosted on AWS-RDS
-Host: rxidds.cqqygklpjkea.us-east-2.rds.amazonaws.com 
-
-Port: 5432
-
-DB name: rxidDS 
+includes CORS and HTTPS functionality
