--- Web Service Endpoint/README.md
@@ -1,3 +1,0 @@
-# DS Backend Module  setup for deployment to AWS
-
-includes CORS and HTTPS functionality
