--- Web Service Endpoint/requirements.txt
@@ -1,20 +1,0 @@
-numpy==1.15.4
-pandas
-flask>=0.12.3
-flask-cors
-flask-sqlalchemy
-python-decouple 
-python-dotenv 
-gunicorn 
-scikit-learn
-scipy
-joblib==0.13.2
-category-encoders
-psycopg2-binary
-sqlalchemy
-boto3
-opencv-python
-selenium
-beautifulsoup4
-urllib3>=1.24.2
-python-dotenv
