+++ requirements.txt
@@ -1,7 +1,6 @@
-<<<<<<< HEAD
 numpy==1.15.4
 pandas
-flask
+flask>=0.12.3
 flask-cors
 flask-sqlalchemy
 python-decouple 
@@ -17,356 +16,5 @@ boto3
 opencv-python
 selenium
 beautifulsoup4
-=======
-absl-py==0.7.0
-aiohttp==3.5.4
-alabaster==0.7.12
-altair==2.2.2
-appdirs==1.4.3
-arviz==0.3.1
-asn1crypto==0.24.0
-astor==0.7.1
-astroid==2.1.0
-astropy==3.1.1
-astunparse==1.6.2
-async-timeout==3.0.1
-atomicwrites==1.3.0
-attrs==18.2.0
-automl-gs==0.2.1
-autopep8==1.4
-babel==2.6.0
-backcall==0.1.0
-basilica==0.2.5
-beautifulsoup4==4.6.3
-bitarray==0.8.3
-bkcharts==0.2
-bleach==3.1.0
-blinker==1.4
-blosc==1.8.1
-bokeh==1.0.4
-boto==2.49.0
-boto3==1.9.123
-botocore==1.12.123
-bottleneck==1.2.1
-bs4==0.0.1
-builtwith==1.3.2
-bz2file==0.98
-cachetools==2.1.0
-cairocffi==0.9.0
-category_encoders==1.3.0
-certifi==2018.8.24
-cffi==1.11.5
-cftime==1.0.3.4
-chardet==3.0.4
-chromelogger==0.4.3
-click==7.0
-cloudpickle==0.7.0
-clyent==1.2.1
-colorama==0.4.1
-colour==0.1.5
-configparser==3.7.3
-console-log==0.2.10
-contextlib2==0.5.5
-cryptography==2.3.1
-cssselect==1.0.3
-cudf==0.6.1
-cupy==5.1.0
-cycler==0.10.0
-cython==0.29.4
-cytoolz==0.9.0.1
-dash==0.41.0
-dash-core-components==0.46.0
-dash-daq==0.1.4
-dash-html-components==0.15.0
-dash-renderer==0.22.0
-dash-table==3.6.0
-dask==1.1.1
-databricks-cli==0.8.4
-datashape==0.5.2
-dautil==0.0.1b1
-deap==1.2.2
-decorator==4.3.2
-distributed==1.25.3
-django==2.1.5
-dnspython==1.16.0
-docutils==0.14
-editdistance==0.4
-entrypoints==0.3
-et_xmlfile==1.0.1
-fastcache==1.0.2
-fastrlock==0.4
-feedfinder2==0.0.4
-feedparser==5.2.1
-filelock==3.0.10
-flask==0.11.1
-flask-caching==1.3.3
-flask-compress==1.4.0
-flask-cors==3.0.7
-flask-sqlalchemy==2.3.2
-flask-wtf==0.14.2
-future==0.16.0
-gast==0.2.2
-gensim==3.7.1
-geographiclib==1.49
-geojson==2.4.1
-geopy==1.17.0
-gevent==1.4.0
-gevent-websocket==0.10.1
-glob2==0.6
-google-api-python-client==1.6.7
-google-auth==1.5.1
-google-images-download==2.5.0
-googlemaps==2.5.1
-greenlet==0.4.15
-grpcio==1.14.1
-gunicorn==19.9.0
-gym==0.7.4
-h5py==2.9.0
-heapdict==1.0.0
-html5lib==1.0.1
-httplib2==0.11.3
-idna==2.8
-idna_ssl==1.1.0
-imageio==2.4.1
-imagesize==1.1.0
-importlib_metadata==0.7
-intel-openmp==2019.0
-ipykernel==5.1.0
-ipython==7.2.0
-ipython_genutils==0.2.0
-ipywidgets==7.4.2
-isort==4.3.4
-itsdangerous==1.1.0
-jdcal==1.4
-jedi==0.13.2
-jeepney==0.4
-jieba3k==0.35.1
-jinja2==2.10
-joblib==0.13.1
-jsonpickle==1.1
-jsonschema==2.6.0
-jupyter==1.0.0
-jupyter-server-proxy==1.0
-jupyter_client==5.2.4
-jupyter_console==6.0.0
-jupyter_contrib_core==0.3.3
-jupyter_contrib_nbextensions==0.5.0
-jupyter_core==4.4.0
-jupyter_highlight_selected_word==0.2.0
-jupyter_latex_envs==1.4.4
-jupyter_nbextensions_configurator==0.4.0
-jupyterlab==0.35.3
-jupyterlab-code-formatter==0.2.0
-jupyterlab-dash==0.1.0a2
-jupyterlab_launcher==0.13.1
-jupyterlab_server==0.2.0
-kaggle==1.5.2
-keras==2.2.4
-keras-applications==1.0.6
-keras-preprocessing==1.0.5
-keyring==17.1.1
-kiwisolver==1.0.1
-landslide==1.1.6
-lazy-object-proxy==1.3.1
-libarchive==0.4.6
-line_profiler==2.1.2
-livereload==2.6.0
-llvmlite==0.27.0
-locket==0.2.0
-lxml==4.3.0
-mako==1.0.7
-markdown==2.6.11
-markupsafe==1.1.1
-more-itertools==4.3.0
-mpc==0.0.3
-mpld3==0.3
-mpmath==1.1.0
-msgpack==0.5.6
-msgpack-python==0.5.6
-multidict==4.5.2
-multipledispatch==0.6.0
-nbconvert==5.3.1
-nbformat==4.4.0
-netcdf4==1.4.2
-networkx==2.2
-newspaper3k==0.2.8
-nltk==3.4
-nose==1.3.7
-notebook==5.7.4
-numba==0.42.0
-numexpr==2.6.9
-numpy==1.16.1
-matplotlib==3.0.2
-mccabe==0.6.1
-memory_profiler==0.55.0
-mistune==0.8.4
-mkl==2019.0
-numpydoc==0.8.0
-nvstrings==0.3.0.post1
-oauth2client==4.1.2
-oauthlib==3.0.1
-odo==0.5.0
-olefile==0.46
-opencv-python==4.1.0.25
-openpyxl==2.5.14
-packaging==19.0
-pandas==0.24.2
-pandas-datareader==0.7.0
-pandoc==1.0.2
-pandocfilters==1.4.2
-paralleldots==3.2.12
-parso==0.3.2
-partd==0.3.9
-path.py==11.5.0
-pathlib-mate==0.0.15
-pathlib2==2.3.3
-patsy==0.5.1
-pdpbox==0.2.0
-pep8==1.7.1
-pexpect==4.6.0
-pickleshare==0.7.5
-pika==0.12.0
-pillow==6.0.0
-pip==19.0.1
-pixiedust==1.1.15
-pixiedust-node==0.2.5
-pkginfo==1.5.0.1
-plotly==3.8.0
-pluggy==0.8.1
-ply==3.11
-prometheus_client==0.5.0
-prompt_toolkit==2.0.8
-protobuf==3.6.1
-psutil==5.5.0
-psycopg2==2.7.5
-ptyprocess==0.6.0
-py==1.7.0
-py-openaq==1.1.0
-pyarrow==0.12.1
-pyasn1==0.4.5
-pyasn1-modules==0.2.3
-pycodestyle==2.5.0
-pycosat==0.6.3
-pycparser==2.19
-pycrypto==2.6.1
-pydrive==1.3.1
-pyflakes==2.1.0
-pyglet==1.3.2
-pygments==2.3.1
-pyjwt==1.7.1
-pylint==2.2.2
-pymodel==1.0
-pymongo==3.7.2
-pyodbc==4.0.25
-pyopenssl==19.0.0
-pyparsing==2.3.1
-pysocks==1.6.8
-pytesseract==0.2.6
-pytest==4.2.0
-pytest-arraydiff==0.3
-pytest-astropy==0.5.0
-pytest-doctestplus==0.2.0
-pytest-openfiles==0.3.2
-pytest-remotedata==0.3.1
-python-dateutil==2.7.5
-python-decouple==3.1
-python-dotenv==0.10.1
-python-slugify==2.0.1
-pytz==2018.9
-pywavelets==1.0.1
-pyyaml==3.13
-pyzmq==17.1.3
-qtawesome==0.5.6
-qtconsole==4.4.3
-qtpy==1.6.0
-readline==6.2.4.1
-readme-renderer==24.0
-redis==3.2.1
-regex==2018.11.22
-requests==2.20.0
-requests-file==1.4.3
-requests-oauthlib==1.2.0
-requests-toolbelt==0.9.1
-retrying==1.3.3
-rope==0.11.0
-rsa==3.4.2
-ruamel_yaml==0.15.46
-scikit-image==0.14.1
-scikit-learn==0.20.2
-scipy==1.2.0
-seaborn==0.9.0
-secretstorage==3.1.0
-selenium==3.141.0
-send2trash==1.5.0
-setuptools==40.7.3
-shap==0.28.5
-simpervisor==0.3
-simplegeneric==0.8.1
-simplejson==3.16.0
-singledispatch==3.4.0.3
-sip==4.19.8
-six==1.12.0
-smart-open==1.8.0
-snappy
-snowballstemmer==1.2.1
-sortedcollections==1.1.2
-sortedcontainers==2.1.0
-soupsieve==1.7.1
-sphinx==1.8.4
-sphinxcontrib-websupport==1.1.0
-spyder==3.3.2
-spyder-kernels==0.4.2
-sqlalchemy==1.2.17
-statsmodels==0.9.0
-stopit==1.1.2
-sympy==1.3
-tabulate==0.8.3
-tblib==1.3.2
-tbselenium==0.3.3
-tensorboard==1.12.2
-tensorflow-hub==0.4.0
-termcolor==1.1.0
-terminado==0.8.1
-testpath==0.4.2
-textgenrnn==1.5.0
-tffm==1.0.1
-theano==1.0.3
-tinysegmenter==0.3
-tldextract==2.2.1
-toolz==0.9.0
-tornado==5.1.1
-tpot==0.9.6
-tqdm==4.29.1
-traitlets==4.3.2
-tweepy==3.5.0
-twine==1.12.1
-typed-ast==1.1.0
-typing==3.6.4
-unicodecsv==0.14.1
-unidecode==1.0.23
-update-checker==0.16
-uritemplate==3.0.0
-urllib3==1.22
-uszipcode==0.2.2
-vega==1.4.0
-vega_datasets==0.5.0
-wcwidth==0.1.7
-webencodings==0.5.1
-werkzeug==0.14.1
-wget==3.2
-wheel==0.32.3
-widgetsnbextension==3.4.2
-wrapt==1.11.1
-wsgigzip==0.1.4
-wtforms==2.2.1
-wurlitzer==1.0.2
-xarray==0.11.0
-xgboost==0.81
-xlrd==1.2.0
-xlsxwriter==1.1.2
-xlwt==1.3.0
-yapf==0.25.0
-yarl==1.3.0
-zict==0.1.3
-zstd==1.3.8.0
->>>>>>> 073ad39da74296166c6663ce4eb236fa50451c83
+urllib3>=1.24.2
+python-dotenv
